requirements.txt
fastmcp>=2.10.6
fastapi>=0.68.0
uvicorn>=0.15.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5 